Some hard coding, initial communication tests
V0.0.1 - Caution I've hard coded various things as I'm tired and
writing code (not a good thing). I'll move the code to the config file
in the next update.

What works: connects to the Blackbean IR controller and puts it into
learn mode, then gets data when a button is hit.

What doesn't work: the IR appears to be different each time the button
is pressed. The signal captured does work for sending but can't work
for using the IR as a trigger.